+++ backend/app/config.py
@@ -18,4 +18,6 @@ class Config:
     # os.environ crashes immediately
     SQLALCHEMY_DATABASE_URI = os.environ["DATABASE_URL"]
     SQLALCHEMY_TRACK_MODIFICATIONS = False
+    # change to os.environ for production
+    STRIPE_WEBHOOK_SECRET = os.getenv("STRIPE_WEBHOOK_SECRET")