+++ backend/app/models/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env/python3
 
-# filename: __init__.py
+# filename: /models/__init__.py
 '''
 descr: make all folders packages
 helper of 'from app import models' or import app.models
@@ -8,12 +8,14 @@ loads/exports models
 
 '''
 from app.models.admin_user import AdminUser
-#from app.models.user import User # [info] user foreignkeyerror
+from app.models.user import User # [info] user foreignkeyerror
 from app.models.customers import Customer
 from app.models.products import Product, ProductVariant
 from app.models.invoice import Invoice, InvoiceItem, InvoiceTax, InvoiceStatus
 from app.models.payments import Payment
-from app.models.orders import Order
+from app.models.orders import Order, OrderItem, OrderStatus
+# cart is the cart.py filename
+from app.models.cart import Cart, CartItem, CartStatus
 
 __all__ = [
     'User',
@@ -26,4 +28,10 @@ __all__ = [
     'InvoiceTax',
     'Payment',
     'Order',
+    'OrderItem',
+    'OderStatus',
+    'Cart',
+    'CartItem',
+    'CartStatus',
+    'User'
 ]